Move overlay_boxes to image utils
